improve setup.sh and unsetup.sh, update their checksums

+++ setup.sh
@@ -1,45 +1,25 @@
 #!/bin/sh
 
-# run this script to set up the werc environment on openbsd
-# http://werc.cat-v.org/
+# Run this script to set up the Werc environment on OpenBSD.
+# Werc's website: http://werc.cat-v.org/
 
-# exit on first error
-set -o errexit
+# ---- variables ----
 
-# check os
-test "$(uname)" != "OpenBSD" && { echo "$0: operating system is not OpenBSD" >&2 ; exit 1 ; }
+# This section contains customizable variables, consider setting their values
+# before running the script.
 
-# check root
-test "$(whoami)" != "root" && { echo "$0: not running as root" >&2 ; exit 1 ; }
-
-# ----   begin   ----
-
-# this section contains customizable variables, consider setting their values before running the script
-
-# the domain of your server
-# an invalid domain may result in an unsuccessful or incomplete installation
+# The domain of your server.
+# An invalid domain may result in an unsuccessful or incomplete installation.
 domain='example.com'
 
-# directory where the httpd chroot environment will be
-# this is ok for most cases; change this only if you know what you're doing
+# The root directory for httpd's chroot environment.
+# The default value is usually fine and it should not be changed unless the
+# change is backed by a valid reason. If unsure, do not change.
 webdir='/var/www'
 
-# ----    end    ----
+# ---- end variables ----
 
-# for security reasons, assign default values if unset or empty
-webdir=${webdir:-"/var/www"}
-domain=${domain:-"example.com"}
-
-# check webdir value
-echo "$webdir" | grep -E '^(/|(/[[:alnum:]._][-[:alnum:]._]*)+)$' >/dev/null
-if [ $? -eq 1 ]
-then
-	echo "$0: invalid chroot directory" >&2
-	exit 1
-fi
-test $webdir = "/" && echo "$0: careful, webdir is root" >&2
-
-# ----   functions   ----
+# ---- functions ----
 
 # check if a directory (or one of the parents of that directory) is an fstab entry marked as "nodev"
 # directories marked as "nodev" cannot contain special devices (e.g. /dev/null)
@@ -69,17 +49,37 @@ fstab_parent() {
 	return 1
 }
 
-# acts like `ln` if possible, otherwise `cp`
+# acts like ln if possible, otherwise cp
 lncp() {
-	ln $* 2>/dev/null || cp $*
+	ln "$@" 2>/dev/null || cp "$@"
 }
 
 # ---- end functions ----
 
+# exit on first error
+set -o errexit
+
+# check os
+test "$(uname)" != "OpenBSD" && { echo "$0: operating system is not OpenBSD" >&2 ; exit 1 ; }
+
+# check root
+test "$(whoami)" != "root" && { echo "$0: not running as root" >&2 ; exit 1 ; }
+
+# default values if unset or empty
+domain=${domain:-"example.com"}
+webdir=${webdir:-"/var/www"}
+
+# check webdir's value
+echo "$webdir" | grep -E '^(/[^[:cntrl:]]+)+$' >/dev/null
+if [ $? -eq 1 ]
+then
+	echo "$0: invalid chroot directory" >&2
+	exit 1
+fi
+
 pkg_add bzip2 plan9port
 p9pdir='/usr/local/plan9'
 
-# download werc into the environment and set it up
 ftp -S dont http://code.9front.org/hg/werc/archive/tip.tar.bz2
 tar xjf tip.tar.bz2 -C $webdir
 rm tip.tar.bz2
@@ -92,13 +92,13 @@ cp -r $webdir/werc/lib $siteroot/_werc
 printf "# congratulations\n\nit works! :)\n" >$siteroot/index.md
 
 # backup current httpd.conf
-if [ -f /etc/httpd.conf ]
+if [ -r /etc/httpd.conf ]
 then
 	cp /etc/httpd.conf /etc/httpd.conf.bk
 	echo "$0: /etc/httpd.conf already exists, it has been copied to /etc/httpd.conf.bk" >&2
 fi
 
-# write new httpd.conf
+# new httpd.conf
 # for some reason, httpd waits until timeout ("connection request timeout") for some files
 echo \
 'server "'$domain'" {
@@ -142,7 +142,7 @@ then
 	newline=`echo "$oldline" | sed 's/nodev//' | sed 's/,,/,/'`
 	oldfile=`cat /etc/fstab`
 	echo "$oldfile" | sed "s!$oldline!$newline!" >/etc/fstab
-	echo "$0: /etc/fstab changed, a reboot is required at the end of the setup process" >&2
+	echo "$0: /etc/fstab has been changed, a reboot is required at the end of the setup process"
 fi
 
 # create devices in $webdir
@@ -156,7 +156,7 @@ cd $p
 mkdir -p "$webdir/tmp"
 chmod 1777 "$webdir/tmp"
 
-# hard-link (or copy) required things into the chroot environment
+# lncp required things into the chroot environment
 mkdir -p $webdir$p9pdir $webdir/usr/libexec $webdir/usr/lib $webdir/bin $webdir$p9pdir/lib
 lncp $p9pdir/rcmain $webdir$p9pdir
 lncp /usr/libexec/ld.so $webdir/usr/libexec
@@ -164,7 +164,7 @@ lncp /usr/lib/lib{m,util,pthread,c,z,expat}.so* $webdir/usr/lib
 lncp /bin/{pwd,mv} $webdir/bin
 lncp $p9pdir/lib/fortunes $webdir$p9pdir/lib
 
-# recursively hard-link (or copy) everyting (including sub-dirs) under $p9pdir/bin into the chroot environment
+# recursively lncp everyting (including sub-dirs) under $p9pdir/bin into the chroot environment
 allbins="$(find $p9pdir/bin -not -type d | sed "s|^$p9pdir/bin/||")"
 for bin in $allbins
 do
@@ -176,7 +176,7 @@ done
 # enable slowcgi and httpd
 rcctl enable slowcgi httpd
 
-echo "$0: setup completed!" >&2
-echo "$0: check prior messages to see if you need to reboot; otherwise, you can start the httpd and slowcgi services" >&2
+echo "$0: setup completed!"
+echo "$0: check prior messages to see if you need to reboot; otherwise, you can start the httpd and slowcgi services"
 
 exit 0
